Relocating I/O adapters of MessageBufferInput/Output. Removing utility methods in MessagePack class
