+++ msgpack-core/src/main/java/org/msgpack/core/buffer/MessageBufferBE.java
@@ -1,4 +1,4 @@
-package org.msgpack.core;
+package org.msgpack.core.buffer;
 
 import java.nio.ByteBuffer;
