+++ msgpack-core/src/main/java/org/msgpack/core/buffer/MessageBufferInput.java
@@ -1,0 +1,28 @@
+package org.msgpack.core.buffer;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.ByteBuffer;
+import java.nio.channels.ReadableByteChannel;
+
+/**
+ * Provides a sequence of MessageBuffers that contains message packed data.
+ */
+public interface MessageBufferInput {
+
+    /**
+     * Get a next buffer to read
+     * @return
+     * @throws IOException
+     */
+    public MessageBuffer next() throws IOException;
+
+    /**
+     * Close this buffer input
+     * @throws IOException
+     */
+    public void close() throws IOException;
+
+}
+
+
